feat: refactor useConfirmation hook with type definitions - Extract type definitions to separate types.ts file - Add UseConfirmationOptions and ConfirmationOperations interfaces - Improve hook structure and maintainability
