+++ frontend/src/hooks/useConfirmation/types.ts
@@ -1,0 +1,8 @@
+export interface UseConfirmationOptions {
+  message?: string;
+  title?: string;
+}
+
+export interface ConfirmationOperations {
+  confirm: (action: () => Promise<void> | void, options?: UseConfirmationOptions) => Promise<void>;
+}
